Bug 1329945 - move heka_message_parser to moztelemetry.heka
