+++ moztelemetry/heka/message_parser.py
